# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "This API handles the management of customer orders",
        "contact": {
            "email": "aloshmohammad2001@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/change-role": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change user role",
                "description": "Update the role of a user. The role will be assigned to the user based on the provided role name.",
                "operationId": "be3368367109d127080eafbe9c212077",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userId",
                                    "role"
                                ],
                                "properties": {
                                    "userId": {
                                        "description": "ID of the user whose role is to be updated",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role": {
                                        "description": "Name of the role to be assigned",
                                        "type": "string",
                                        "example": "Admins"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User role updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found or User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}